+++ src/main/java/dnaaaaahtac/wooriforei/domain/openapi/dto/information/InformationResponseDto.java
@@ -1,4 +1,4 @@
-package dnaaaaahtac.wooriforei.domain.openapi.dto;
+package dnaaaaahtac.wooriforei.domain.openapi.dto.information;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
@@ -8,7 +8,7 @@ import lombok.NoArgsConstructor;
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-public class InformationResponseWrapper {
+public class InformationResponseDto {
 
     @JsonProperty("TbTourInformation")
     private InformationResponseVo tbTourInformation;
